requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
sentence-transformers
chromadb
google-generativeai
langchain
